Move accordion to primitives folder
