+++ reflex/components/radix/themes/__init__.py
@@ -6,10 +6,3 @@ from .typography import *
 
 theme = Theme.create
 theme_panel = ThemePanel.create
-
-accordion_root = AccordionRoot.create
-accordion_item = AccordionItem.create
-accordion_trigger = AccordionTrigger.create
-accordion_content = AccordionContent.create
-accordion_header = AccordionHeader.create
-chevron_down_icon = ChevronDownIcon.create
